Add authentication

+++ controllers/docsCtrl.js
@@ -9,20 +9,50 @@ async function create(req, res){
 
     try {
         //
-        let create_obj = {
-            empCode: req.body.empCode,
-            documentId: req.body.documentId,
-            name: req.body.name,
-            type: req.body.type,
-            firebaseLink: req.body.firebaseLink
-        };
-
-        let documentsCreated = await db.public.docs.create(create_obj);
-
-        res.status(200).json({
-            success: true,
-            docs: documentsCreated
-        });
+        var authTOKEN = req.header('X-AUTH-TOKEN');
+        if(authTOKEN == "" || authTOKEN == null) {
+            res.status(500).json({
+                success: false,
+                error: {
+                    message: "Token not passed"
+                }
+            });
+        }
+        try{
+            var user = utilities.decryptJWTWithToken(authTOKEN)
+        }    
+        catch{
+            res.status(500).json({
+                success: false,
+                error: {
+                    message: "invalid Token"
+                }
+            });    
+        }
+        if(user) {
+            let create_obj = {
+                empCode: req.body.empCode,
+                documentId: req.body.documentId,
+                name: req.body.name,
+                type: req.body.type,
+                firebaseLink: req.body.firebaseLink
+            };
+    
+            let documentsCreated = await db.public.docs.create(create_obj);
+    
+            res.status(200).json({
+                success: true,
+                docs: documentsCreated
+            });    
+        }
+        else {
+            res.status(500).json({
+                success: false,
+                error: {
+                    message: "Token not found"
+                }
+            });
+        }
 
     } catch(err) {
         console.log(err);
@@ -41,21 +71,52 @@ async function get(req, res){
 
     try {
         //
-        let query = {};
-
-        if(req.query.empCode){
-            query.empCode = req.query.empCode;
+        var authTOKEN = req.header('X-AUTH-TOKEN');
+        if(authTOKEN == "" || authTOKEN == null) {
+            res.status(500).json({
+                success: false,
+                error: {
+                    message: "Token not passed"
+                }
+            });
         }
+        try{
+            var user = utilities.decryptJWTWithToken(authTOKEN)
+        }    
+        catch{
+            res.status(500).json({
+                success: false,
+                error: {
+                    message: "invalid Token"
+                }
+            });    
+        }
+        if(user) {
+            let query = {};
 
-        let values = await db.public.docs.findAll({
-            where: query
-        })
+            if(req.query.empCode){
+                query.empCode = req.query.empCode;
+            }
+
+            let values = await db.public.docs.findAll({
+                where: query
+            })
 
 
-        res.status(200).json({
-            success: true,
-            docs: values
-        });
+            res.status(200).json({
+                success: true,
+                docs: values
+            });
+        }
+        else {
+            res.status(500).json({
+                success: false,
+                error: {
+                    message: "Token not found"
+                }
+            });
+        }
+        
 
     } catch (err) {
         console.log(err);
@@ -74,26 +135,58 @@ async function destroy(req, res){
 
     try {
         //
-        let query = {};
+        var authTOKEN = req.header('X-AUTH-TOKEN');
+        if(authTOKEN == "" || authTOKEN == null) {
+            res.status(500).json({
+                success: false,
+                error: {
+                    message: "Token not passed"
+                }
+            });
+        }
+        try{
+            var user = utilities.decryptJWTWithToken(authTOKEN)
+        }    
+        catch{
+            res.status(500).json({
+                success: false,
+                error: {
+                    message: "invalid Token"
+                }
+            });    
+        }
+        if(user) {
+            let query = {};
 
-        if(req.query.documentId){
-            query.documentId = req.query.documentId;
+            if(req.query.documentId){
+                query.documentId = req.query.documentId;
+            }
+            
+            let values = await db.public.docs.destroy({
+                where: query
+            })
+            if(values==0){
+                res.status(200).json({
+                    error:"document not found"
+                });
+            }else{
+                res.status(200).json({
+                    success: true,
+                    docs: values,
+                    deleted: true
+                });
+            }
         }
-           
-        let values = await db.public.docs.destroy({
-            where: query
-        })
-        if(values==0){
-            res.status(200).json({
-                error:"document not found"
-            });
-        }else{
-            res.status(200).json({
-                success: true,
-                docs: values,
-                deleted: true
+        else {
+            res.status(500).json({
+                success: false,
+                error: {
+                    message: "Token not found"
+                }
             });
         }
+    
+        
 
 
     } catch (err) {
